# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelForge Local API",
    "version": "0.1.0-alpha.3",
    "description": "Project lifecycle, formal local Web workspace, durable events and engine adapter."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:48090"
    }
  ],
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "streamEvents",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Bad"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "last_event_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "project",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      }
    },
...
    "/api/projects": {
      "get": {
        "operationId": "listProjects",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Bad"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "name": "archived",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "all"
              ]
            }
          },
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ]
      },
      "post": {
        "operationId": "createProject",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Bad"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/truth/events": {
      "get": {
        "operationId": "listTruthEvents",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "through_chapter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "subject_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "predicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Truth event page in stable sequence order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthEventPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "appendTruthEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TruthAppendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Truth event appended atomically",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthAppendResult"
                }
              }
            }
          },
          "200": {
            "description": "Exact response replayed from the persistent idempotency ledger",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthAppendResult"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        }
      },
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      }
    },
    "responses": {
      "Bad": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "TooLarge": {
        "description": "Request body too large",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Internal": {
        "description": "Internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "details",
          "retryable",
          "trace_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "retryable": {
            "type": "boolean"
          },
          "trace_id": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "ProjectSummary": {
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "archived",
          "current_chapter",
          "completed_chapters",
          "total_chapters",
          "total_words",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "archived": {
            "type": "boolean"
          },
          "phase": {
            "type": "string"
          },
          "current_chapter": {
            "type": "integer",
            "minimum": 0
          },
          "completed_chapters": {
            "type": "integer",
            "minimum": 0
          },
          "total_chapters": {
            "type": "integer",
            "minimum": 0
          },
          "total_words": {
            "type": "integer",
            "minimum": 0
          },
          "current_volume": {
            "type": "integer",
            "minimum": 0
          },
          "current_arc": {
            "type": "integer",
            "minimum": 0
          },
          "format_version": {
            "type": "integer",
            "minimum": 1
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProjectDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProjectSummary"
          },
          {
            "type": "object",
            "properties": {
              "synopsis": {
                "type": "string"
              },
              "genre": {
                "type": "string"
              },
              "language": {
                "type": "string"
              },
              "target_words": {
                "type": "integer",
                "minimum": 0
              },
              "words_per_chapter": {
                "type": "integer",
                "minimum": 0
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "archived_at": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "source_format": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ProjectList": {
        "type": "object",
        "required": [
          "projects",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "next_offset": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "CreateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "synopsis": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "target_words": {
            "type": "integer",
            "minimum": 0
          },
          "target_chapters": {
            "type": "integer",
            "minimum": 0
          },
          "words_per_chapter": {
            "type": "integer",
            "minimum": 0
          },
          "slug": {
            "type": "string",
            "maxLength": 80
          },
          "import_path": {
            "type": "string"
          }
        },
        "anyOf": [
          {
            "required": [
              "title"
            ]
          },
          {
            "required": [
              "import_path"
            ]
          }
        ]
      },
...
      "TruthSource": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "id",
          "chapter",
          "version"
        ],
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "chapter": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "extractor": {
            "type": "string",
            "maxLength": 200
          },
          "confirmed_by": {
            "type": "string",
            "maxLength": 200
          },
          "excerpt": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
      "TruthEventInput": {
        "type": "object",
        "required": [
          "kind",
          "subject_type",
          "subject_id",
          "predicate",
          "value",
          "valid_from_chapter",
          "known_from_chapter",
          "authority",
          "confidence",
          "source"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "assert",
              "supersede",
              "retract"
            ]
          },
          "subject_type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
          },
          "subject_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "predicate": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
          },
          "value": {},
          "valid_from_chapter": {
            "type": "integer",
            "minimum": 0
          },
          "valid_to_chapter": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "known_from_chapter": {
            "type": "integer",
            "minimum": 0
          },
          "known_to_chapter": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "authority": {
            "type": "string",
            "enum": [
              "llm_suggestion",
              "story_compass",
              "volume_plan",
              "arc_plan",
              "chapter_plan",
              "generated_final",
              "human_final"
            ]
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "source": {
            "$ref": "#/components/schemas/TruthSource"
          },
          "supersedes_event_id": {
            "type": "string",
            "maxLength": 200
          }
        },
        "additionalProperties": false
      },
      "TruthAppendRequest": {
        "type": "object",
        "required": [
          "project_id",
          "event"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "event": {
            "$ref": "#/components/schemas/TruthEventInput"
          }
        },
        "additionalProperties": false
      },
      "TruthEvent": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TruthEventInput"
          },
          {
            "type": "object",
            "required": [
              "sequence",
              "id",
              "created_at",
              "checksum"
            ],
            "properties": {
              "sequence": {
                "type": "integer"
              },
              "id": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "checksum": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "TruthAppendResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "event",
          "replayed"
        ],
        "properties": {
          "event": {
            "$ref": "#/components/schemas/TruthEvent"
          },
          "replayed": {
            "type": "boolean"
          }
        }
      },
      "TruthEventPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "events",
          "after_sequence",
          "limit"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruthEvent"
            }
          },
          "after_sequence": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "next_sequence": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
    }
  }
}